+++ oap-server/server-storage-plugin/storage-banyandb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/banyandb/stream/BanyanDBRecordDAO.java
@@ -44,7 +44,7 @@ public class BanyanDBRecordDAO implements IRecordDAO {
         StreamWrite streamWrite = new StreamWrite(
             schema.getMetadata().getGroup(), // group name
             model.getName(), // index-name
-            record.id() // identity
+            record.id().build() // identity
         ); // set timestamp inside `BanyanDBConverter.StreamToStorage`
         Convert2Storage<StreamWrite> convert2Storage = new BanyanDBConverter.StreamToStorage(schema, streamWrite);
         storageBuilder.entity2Storage(record, convert2Storage);